extract LayoutButton and DisabledButton from Button.js
