+++ unlock-app/src/components/interface/buttons/layout/Close.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Svg from '../../svg'
-import { LayoutButton } from '../Button'
+import LayoutButton from '../LayoutButton'
 
 const Close = props => (
   <LayoutButton {...props}>
