+++ unlock-app/src/components/interface/buttons/layout/Github.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Svg from '../../svg'
-import { LayoutButton } from '../Button'
+import LayoutButton from '../LayoutButton'
 
 const Github = props => (
   <LayoutButton
