+++ unlock-app/src/components/interface/buttons/layout/Newsletter.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Svg from '../../svg'
-import { LayoutButton } from '../Button'
+import LayoutButton from '../LayoutButton'
 
 const Newsletter = props => (
   <LayoutButton href="/newsletter" label="Newsletter" {...props}>
