+++ unlock-app/src/components/interface/buttons/layout/Twitter.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Svg from '../../svg'
-import { LayoutButton } from '../Button'
+import LayoutButton from '../LayoutButton'
 
 const Twitter = props => (
   <LayoutButton href="/twitter" label="Twitter" {...props}>
